pyproject.toml
[tool.poetry]
name = "ml-zoomcamp"
version = "0.1.0"
description = ""
authors = ["JChien <jeffchien13@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.23.1"
pandas = "^1.4.3"
scikit-learn = "^1.1.1"
tensorflow-macos = { version = "^2.9.2", platform = "darwin" }
tensorflow-metal = { version = "^0.5.0", platform = "darwin" }
xgboost = "^1.6.1"
onnxruntime = "^1.12.1"
Flask = "^2.2.2"
grpcio = "^1.48.0"
protobuf = "3.19.4"
gunicorn = "^20.1.0"
certifi = "*"

[tool.poetry.dev-dependencies]
pyarrow = "^8.0.0"
kaggle = "^1.5.12"
graphviz = "^0.20.1"
s3fs = "^2022.7.1"
black = "^22.6.0"
onnxmltools = "^1.11.1"
seaborn = "^0.11.2"
jupyter = "^1.0.0"
matplotlib = "^3.5.2"
grpcio-tools = "^1.48.0"
requests = "^2.28.1"
watchdog = { extras = ["watchmedo"], version = "^2.1.9" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
